bam/bai: Decouple the index from the record reader and writer
The index read its integers with `io::reader::num` and wrote them with
`io::writer::num`, which are private helpers of the record reader and
writer. That is the wrong direction: the index knows nothing about
records, and the BAI format is defined independently of them.

It now has its own `bai::io::num`. `read_u64_le` and `write_u64_le`
existed in the record helpers only for the index and are removed, which
is what confirms where they belonged.
