+++ noodles-bam/src/bai/io/writer/index/reference_sequences/bins.rs
@@ -7,7 +7,7 @@ use noodles_csi::binning_index::index::reference_sequence::{Bin, Metadata};
 
 use self::chunks::write_chunks;
 use super::write_metadata;
-use crate::io::writer::num::write_u32_le;
+use crate::bai::io::num::write_u32_le;
 
 pub(super) fn write_bins<W>(
     writer: &mut W,
