+++ noodles-bam/src/io/writer/num.rs
@@ -15,11 +15,3 @@ where
     let buf = n.to_le_bytes();
     writer.write_all(&buf)
 }
-
-pub(crate) fn write_u64_le<W>(writer: &mut W, n: u64) -> io::Result<()>
-where
-    W: Write,
-{
-    let buf = n.to_le_bytes();
-    writer.write_all(&buf)
-}
